Huge Update
This is a merger of changes made in the last 4 months.

-Converted the data file from pickle format to an easy-to-read json
format.
-Added the random stripclub generator module
-Added the Urban Dictionary lookup module
-Updated the vocabulary data file (this is still in pickle format, and
may get changed over to json soon)
-Disabled dynamic reloading functionality (due to my lack of linux
knowledge and being unable to fix the crashes hat occur on the bot's
server whenever I try to re-enable the reloader module)
-The bot recognizes "BOSS" as whomever is auth/+x'd on quakenet to the
username set in that variable. (ie: tizenwork)
-AUTH variable needs to be loaded from an outside file (which needs to
be added to .gitignore), so I don't have to remember to clear it every
time I update github.
-Lots of new commands added to trpbot_commands.py